test(checkout): migrate series-pass notification tests to reserve/session; split cleanup tests for file length (#632)
